Move declaration files to src/types/declarations
