+++ src/types/declarations/cmd.d.ts
